console: Restructure folders in managed gateway
