+++ pkg/webui/console/containers/gateway-managed-gateway/shared/show-profiles-select.js
