+++ pkg/webui/console/containers/gateway-managed-gateway/shared/utils.js
